requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-background-tasks==1.2.5
django-braces==1.15.0
django-ckeditor==6.4.2
django-compat==1.0.15
django-filter==21.1
django-iranian-cities==1.0.2
django-jalali==6.0.0
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-model-utils==4.2.0
django-star-ratings==0.9.2
django-taggit==2.1.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
extensions==0.4
future==0.18.2
ghasedak==0.1.7
idna==3.3
jdatetime==4.1.0
Pillow==9.1.0
pytz==2022.1
requests==2.27.1
six==1.16.0
sorl-thumbnail==12.8.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
swapper==1.3.0
urllib3==1.26.9
